+++ packages/stentor-guards/src/index.ts
@@ -1,15 +1,19 @@
 /*! Copyright (c) 2020, XAPPmedia */
 export * from "./isActionable";
 export * from "./isActiveWithin";
+export * from "./isCompilablePath";
 export * from "./isContextual";
+export * from "./isExecutablePath";
 export * from "./isFirstTime";
 export * from "./isGlobalHandler";
 export * from "./isHandler";
 export * from "./isHaveNotSeenWithin";
+export * from "./isHistoricalPath";
 export * from "./isIntent";
 export * from "./isLastActive";
 export * from "./isLastActiveResponse";
 export * from "./isLinkoutSuggestion";
+export * from "./isPreviousHandlerPath";
 export * from "./isSchedulableResponse";
 export * from "./isScheduled";
 export * from "./isTimeContextual";
