+++ packages/stentor-handler/src/Path/compilePreviousHandlerPath.ts
@@ -1,9 +1,9 @@
 /*! Copyright (c) 2019, XAPPmedia */
+import { isPreviousHandlerPath } from "stentor-guards";
 import { Context, ExecutablePath, PreviousHandlerPath, Request } from "stentor-models";
 import { keyFromRequest } from "stentor-request";
 import { findValueForKey } from "stentor-utils";
 import { determinePath } from "./determinePath";
-import { isPreviousHandlerPath } from "./Guards";
 
 export function compilePreviousHandlerPath(
     previousHandlerPath: PreviousHandlerPath,
